+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #1990 Fix items not filtered by Worksheet Template's method in Add analyses
+- #1887 Fix instruments not filtered by method in Worksheet Template edit view
 - #1977 Add custom worksheet layout feature
 - #1988 Multivalue support for interim fields
 - #1987 Fix: ProfileKey shown in Profiles table